+++ gdsfactory/components/switch_tree.py
@@ -20,7 +20,7 @@ from gdsfactory.components.mmi2x2 import mmi2x2
 from gdsfactory.components.mzi_phase_shifter import mzi_phase_shifter
 from gdsfactory.components.splitter_tree import splitter_tree
 
-mzi = gf.partial(mzi_phase_shifter, combiner=mmi2x2)
+mzi = gf.partial(mzi_phase_shifter, combiner=mmi2x2, delta_length=0)
 
 
 switch_tree = gf.partial(splitter_tree, coupler=mzi, dx=500, dy=100)