+++ gdsfactory/components/via_stack_with_offset.py
@@ -6,34 +6,33 @@ import gdsfactory as gf
 from gdsfactory.component import Component
 from gdsfactory.components.via import via1
 from gdsfactory.tech import LAYER
-from gdsfactory.types import ComponentFactory, Layer
+from gdsfactory.types import ComponentOrFactory, Layers
 
 
 @gf.cell
 def via_stack_with_offset(
-    layer_via_width_height_offset: Tuple[
-        Tuple[Layer, Optional[ComponentFactory], Tuple[float, float], float], ...
-    ] = (
-        (LAYER.Ppp, None, (10, 10), 0),
-        (LAYER.M1, via1, (10, 10), 0),
-    ),
+    layers: Layers = (LAYER.Ppp, LAYER.M1),
+    sizes: Tuple[Tuple[float, float], ...] = ((10, 10), (10, 10)),
+    vias: Tuple[Optional[ComponentOrFactory], ...] = (None, via1),
+    offsets: Optional[Tuple[float, ...]] = None,
     port_orientation: int = 180,
 ) -> Component:
     """Rectangular transition thru metal layers with offset between layers
 
     Args:
-        layer_via_width_height_offset:
-            layer:
-            via: factory
-            size:
-            offset: for next layer
+        layers:
+        vias: factory for via or None for no via
+        sizes:
+        offsets: for next layer
         port_orientation: 180: W0, 0: E0, 90: N0, 270: S0
     """
     c = Component()
     x0 = x1 = y0 = y1 = 0
 
+    offsets = offsets or [0] * len(layers)
+
     i = 0
-    for i, (layer, via, size, offset) in enumerate(layer_via_width_height_offset):
+    for i, (layer, via, size, offset) in enumerate(zip(layers, vias, sizes, offsets)):
         width, height = size
         x0 = -width / 2
         x1 = +width / 2
@@ -42,15 +41,13 @@ def via_stack_with_offset(
         c.add_polygon(rect_pts, layer=layer)
 
         if via:
-            via = via()
-            w = via.info["width"]
-            h = via.info["height"]
-            g = via.info["enclosure"]
-            pitch_x = via.info["pitch_x"]
-            pitch_y = via.info["pitch_y"]
+            via = via() if callable(via) else via
+            w, h = via.info["size"]
+            enclosure = via.info["enclosure"]
+            pitch_x, pitch_y = via.info["spacing"]
 
-            nb_vias_x = (width - w - 2 * g) / pitch_x + 1
-            nb_vias_y = (height - h - 2 * g) / pitch_y + 1
+            nb_vias_x = (width - w - 2 * enclosure) / pitch_x + 1
+            nb_vias_y = (height - h - 2 * enclosure) / pitch_y + 1
 
             nb_vias_x = int(floor(nb_vias_x)) or 1
             nb_vias_y = int(floor(nb_vias_y)) or 1
@@ -61,9 +58,11 @@ def via_stack_with_offset(
             x00 = x0 + cw + w / 2
             y00 = y0 + ch + h / 2 + offset
 
-            for i in range(nb_vias_x):
-                for j in range(nb_vias_y):
-                    c.add(via.ref(position=(x00 + i * pitch_x, y00 + j * pitch_y)))
+            ref = c.add_array(
+                via, columns=nb_vias_x, rows=nb_vias_y, spacing=(pitch_x, pitch_y)
+            )
+            ref.move((x00, y00))
+
         y0 += offset
         y1 = y0 + height
         rect_pts = [(x0, y0), (x1, y0), (x1, y1), (x0, y1)]
@@ -86,29 +85,9 @@ def via_stack_with_offset(
 
 
 if __name__ == "__main__":
-
-    # c = via_stack_with_offset(
-    #     layer_via_width_height_offset=(
-    #         (LAYER.Ppp, gf.components.via1, (10, 10), 0),
-    #         (LAYER.M1, None, (10, 10), 10),
-    #     )
-    # )
-    # c = via_stack_with_offset(
-    #     layer_via_width_height_offset=(
-    #         (LAYER.Ppp, gf.components.via1, (10, 10), 10),
-    #         (LAYER.M1, gf.components.via2, (10, 10), 0),
-    #         (LAYER.M2, None, (10, 10), 10),
-    #     )
-    # )
-    # c = via_stack_with_offset(
-    #     layer_via_width_height_offset=(
-    #         (LAYER.Ppp, gf.components.via1, (5, 10), 0),
-    #         (LAYER.M1, gf.components.via2, (5, 10), 10),
-    #         (LAYER.M2, gf.components.via3, (5, 10), 0),
-    #         # (LAYER.M3, None, 5, 10, 0),
-    #     )
-    # )
-    # c.pprint
-    c = via_stack_with_offset()
-    # print(c)
+    c = via_stack_with_offset(
+        layers=(LAYER.SLAB90, LAYER.M1),
+        sizes=((20, 10), (20, 10)),
+        vias=(via1(size=(18, 2), spacing=(5, 5)), None),
+    )
     c.show()